+++ Parser.php
@@ -27,6 +27,7 @@ class Parser
     public const BLOCK_SCALAR_HEADER_PATTERN = '(?P<separator>\||>)(?P<modifiers>\+|\-|\d+|\+\d+|\-\d+|\d+\+|\d+\-)?(?P<comments> +#.*)?';
     public const REFERENCE_PATTERN = '#^&(?P<ref>[^ ]++) *+(?P<value>.*)#u';
     public const DEFAULT_MAX_NESTING_LEVEL = 128;
+    public const DEFAULT_MAX_ALIASES_FOR_COLLECTIONS = 128;
 
     private $filename;
     private $offset = 0;
@@ -41,13 +42,18 @@ class Parser
     private $refsBeingParsed = [];
     private $state;
 
-    public function __construct(int $maxNestingLevel = self::DEFAULT_MAX_NESTING_LEVEL)
+    public function __construct(int $maxNestingLevel = self::DEFAULT_MAX_NESTING_LEVEL, int $maxAliasesForCollections = self::DEFAULT_MAX_ALIASES_FOR_COLLECTIONS)
     {
         if ($maxNestingLevel < 1) {
             throw new \InvalidArgumentException('The maximum nesting depth must be greater than 0.');
         }
 
+        if ($maxAliasesForCollections < 0) {
+            throw new \InvalidArgumentException('The maximum number of collection aliases must be greater than or equal to 0.');
+        }
+
         $this->getState()->maxNestingLevel = $maxNestingLevel;
+        $this->getState()->maxAliasesForCollections = $maxAliasesForCollections;
     }
 
     /**
@@ -98,6 +104,7 @@ class Parser
         $this->refs = [];
         $state = $this->getState();
         $state->reset();
+        $state->aliasesEnabled = 0 === (Yaml::PARSE_EXCEPTION_ON_ALIAS & $flags);
 
         $mbEncoding = null;
 
@@ -273,6 +280,8 @@ class Parser
                         }
 
                         $refValue = $this->refs[$refName];
+
+                        $this->getState()->countAlias($refValue, $this->getRealCurrentLineNb() + 1, $this->currentLine, $this->filename);
 
                         if (Yaml::PARSE_OBJECT_FOR_MAP & $flags && $refValue instanceof \stdClass) {
                             $refValue = (array) $refValue;
@@ -759,6 +768,8 @@ class Parser
 
                 throw new ParseException(sprintf('Reference "%s" does not exist.', $value), $this->currentLineNb + 1, $this->currentLine, $this->filename);
             }
+
+            $this->getState()->countAlias($this->refs[$value], $this->getRealCurrentLineNb() + 1, $this->currentLine, $this->filename);
 
             return $this->refs[$value];
         }