+++ Tests/ParserTest.php
@@ -2967,6 +2967,96 @@ YAML;
         $this->parser->parse($yaml);
     }
 
+    public function testParseRejectsDocumentsThatExceedTheConfiguredCollectionAliasLimit()
+    {
+        $parser = new Parser(Parser::DEFAULT_MAX_NESTING_LEVEL, 5);
+
+        $this->expectException(ParseException::class);
+        $this->expectExceptionMessage('Maximum number of collection aliases');
+
+        $yaml = <<<YAML
+a0: &a0 [foo]
+a1: &a1 [*a0, *a0, *a0]
+payload: [*a1, *a1, *a1]
+YAML;
+
+        $parser->parse($yaml);
+    }
+
+    public function testParseCountsTaggedValueCollectionAliases()
+    {
+        $parser = new Parser(Parser::DEFAULT_MAX_NESTING_LEVEL, 2);
+
+        $this->expectException(ParseException::class);
+        $this->expectExceptionMessage('Maximum number of collection aliases');
+
+        $yaml = <<<YAML
+a: &a !my_tag [foo, bar]
+b: *a
+c: *a
+d: *a
+YAML;
+
+        $parser->parse($yaml, Yaml::PARSE_CUSTOM_TAGS);
+    }
+
+    public function testParseAcceptsScalarAliasesWithoutCountingThem()
+    {
+        $parser = new Parser(Parser::DEFAULT_MAX_NESTING_LEVEL, 1);
+
+        $yaml = <<<YAML
+anchor: &val scalar_value
+a: *val
+b: *val
+c: *val
+d: *val
+YAML;
+
+        $result = $parser->parse($yaml);
+
+        $this->assertSame('scalar_value', $result['a']);
+        $this->assertSame('scalar_value', $result['d']);
+    }
+
+    public function testParseAcceptsLargeCollectionAliasedOnce()
+    {
+        $items = implode(', ', array_map(function ($i) { return "item$i"; }, range(1, 500)));
+        $yaml = <<<YAML
+defaults: &defaults [$items]
+a: *defaults
+YAML;
+
+        $result = $this->parser->parse($yaml);
+
+        $this->assertCount(500, $result['a']);
+    }
+
+    public function testParseRejectsBlockAliasesWhenDisabled()
+    {
+        $this->expectException(ParseException::class);
+        $this->expectExceptionMessage('Aliases are disabled');
+
+        $yaml = <<<YAML
+defaults: &defaults [foo, bar]
+a: *defaults
+YAML;
+
+        $this->parser->parse($yaml, Yaml::PARSE_EXCEPTION_ON_ALIAS);
+    }
+
+    public function testParseRejectsInlineAliasesWhenDisabled()
+    {
+        $this->expectException(ParseException::class);
+        $this->expectExceptionMessage('Aliases are disabled');
+
+        $yaml = <<<YAML
+defaults: &defaults [foo, bar]
+a: [*defaults]
+YAML;
+
+        $this->parser->parse($yaml, Yaml::PARSE_EXCEPTION_ON_ALIAS);
+    }
+
     private function assertSameData($expected, $actual)
     {
         $this->assertEquals($expected, $actual);