+++ Tests/YamlTest.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Yaml\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Yaml\Exception\ParseException;
+use Symfony\Component\Yaml\Parser;
 use Symfony\Component\Yaml\Yaml;
 
 class YamlTest extends TestCase
@@ -47,5 +48,25 @@ class YamlTest extends TestCase
         $this->expectExceptionMessage('Maximum nesting depth of 2 exceeded');
 
         Yaml::parse($yaml, 0, 2);
+    }
+
+    public function testParseFileAllowsConfiguringTheMaximumCollectionAliasCount()
+    {
+        $file = tempnam(sys_get_temp_dir(), 'yaml_');
+
+        file_put_contents($file, <<<YAML
+defaults: &defaults [foo, bar]
+copy: *defaults
+YAML
+);
+
+        $this->expectException(ParseException::class);
+        $this->expectExceptionMessage('Maximum number of collection aliases (0) exceeded.');
+
+        try {
+            Yaml::parseFile($file, 0, Parser::DEFAULT_MAX_NESTING_LEVEL, 0);
+        } finally {
+            unlink($file);
+        }
     }
 }