+++ Yaml.php
@@ -34,6 +34,7 @@ class Yaml
     public const PARSE_CUSTOM_TAGS = 512;
     public const DUMP_EMPTY_ARRAY_AS_SEQUENCE = 1024;
     public const DUMP_NULL_AS_TILDE = 2048;
+    public const PARSE_EXCEPTION_ON_ALIAS = 8192;
 
     /**
      * Parses a YAML file into a PHP value.
@@ -43,17 +44,18 @@ class Yaml
      *     $array = Yaml::parseFile('config.yml');
      *     print_r($array);
      *
-     * @param string $filename        The path to the YAML file to be parsed
-     * @param int    $flags           A bit field of PARSE_* constants to customize the YAML parser behavior
-     * @param int    $maxNestingLevel The maximum nesting depth for nested YAML blocks
+     * @param string $filename                 The path to the YAML file to be parsed
+     * @param int    $flags                    A bit field of PARSE_* constants to customize the YAML parser behavior
+     * @param int    $maxNestingLevel          The maximum nesting depth for nested YAML blocks
+     * @param int    $maxAliasesForCollections The maximum number of collection aliases to resolve
      *
      * @return mixed
      *
      * @throws ParseException If the file could not be read or the YAML is not valid
      */
-    public static function parseFile(string $filename, int $flags = 0, int $maxNestingLevel = Parser::DEFAULT_MAX_NESTING_LEVEL)
+    public static function parseFile(string $filename, int $flags = 0, int $maxNestingLevel = Parser::DEFAULT_MAX_NESTING_LEVEL, int $maxAliasesForCollections = Parser::DEFAULT_MAX_ALIASES_FOR_COLLECTIONS)
     {
-        $yaml = new Parser($maxNestingLevel);
+        $yaml = new Parser($maxNestingLevel, $maxAliasesForCollections);
 
         return $yaml->parseFile($filename, $flags);
     }
@@ -67,17 +69,18 @@ class Yaml
      *   print_r($array);
      *  </code>
      *
-     * @param string $input           A string containing YAML
-     * @param int    $flags           A bit field of PARSE_* constants to customize the YAML parser behavior
-     * @param int    $maxNestingLevel The maximum nesting depth for nested YAML blocks
+     * @param string $input                    A string containing YAML
+     * @param int    $flags                    A bit field of PARSE_* constants to customize the YAML parser behavior
+     * @param int    $maxNestingLevel          The maximum nesting depth for nested YAML blocks
+     * @param int    $maxAliasesForCollections The maximum number of collection aliases to resolve
      *
      * @return mixed
      *
      * @throws ParseException If the YAML is not valid
      */
-    public static function parse(string $input, int $flags = 0, int $maxNestingLevel = Parser::DEFAULT_MAX_NESTING_LEVEL)
+    public static function parse(string $input, int $flags = 0, int $maxNestingLevel = Parser::DEFAULT_MAX_NESTING_LEVEL, int $maxAliasesForCollections = Parser::DEFAULT_MAX_ALIASES_FOR_COLLECTIONS)
     {
-        $yaml = new Parser($maxNestingLevel);
+        $yaml = new Parser($maxNestingLevel, $maxAliasesForCollections);
 
         return $yaml->parse($input, $flags);
     }